Move duplicated code in wrappers to common abstract base class
